Improving coverage.

+++ segmentstore/server/src/main/java/io/pravega/segmentstore/server/DebugSegmentContainer.java
@@ -11,5 +11,5 @@ package io.pravega.segmentstore.server;
 import java.util.concurrent.CompletableFuture;
 
 public interface DebugSegmentContainer extends SegmentContainer {
-    CompletableFuture<Void> createStreamSegment(String streamSegmentName, int length, boolean isSealed);
+    CompletableFuture<Void> createStreamSegment(String streamSegmentName, long length, boolean isSealed);
 }
