+++ segmentstore/server/src/main/java/io/pravega/segmentstore/server/containers/DebugStreamSegmentContainer.java
@@ -62,7 +62,7 @@ public class DebugStreamSegmentContainer extends StreamSegmentContainer implemen
      * @return                          A newly created segment.
      */
     @Override
-    public CompletableFuture<Void> createStreamSegment(String streamSegmentName, int length, boolean isSealed
+    public CompletableFuture<Void> createStreamSegment(String streamSegmentName, long length, boolean isSealed
             /*TODO: pass in generic params */) {
         if (log.isDebugEnabled()) {
             log.debug("createStreamSegment called for {}", streamSegmentName);
